fix: widen 429 detection in classify_tool_output + clippy nit
"HTTP 429 Too Many Requests" was classified as Unknown because
the 429 guard required the word "rate". Add "too many requests",
"http 429", and "status 429" patterns so the standard HTTP
reason phrase is recognised.

Also remove unnecessary i64 cast in test stub (clippy).

+++ rust/crates/runtime/src/lesson_extractor.rs
@@ -818,7 +818,7 @@ mod tests {
             let mut guard = self.lessons.lock().unwrap();
             if let Some(l) = guard.iter_mut().find(|l| l.id == lesson_id) {
                 l.hit_count += 1;
-                Ok(l.hit_count as i64)
+                Ok(l.hit_count)
             } else {
                 Err(sqlx::Error::RowNotFound)
             }
